update github deps
