+++ packages/api/package.json
@@ -12,7 +12,7 @@
     "@date-fns/tz": "^1.2.0",
     "date-fns": "4.1.0",
     "meetup-shared": "workspace:*",
-    "octokit": "^3.1.0",
+    "octokit": "^5.0.3",
     "valibot": "^1.1.0"
   },
   "devDependencies": {
