+++ packages/site/src/pages/submit.astro
@@ -548,7 +548,6 @@ Stable diffusion for the win. For images and animations. But what is facebook do
     }
 
     const json = await res.value.json();
-    console.log(json);
 
     window.location.assign(
       `/success?githubLink=${encodeURIComponent(json.issueUrl)}`,
